requirements/prod.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/prod.in
#
certifi==2022.5.18.1
    # via requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.12
    # via requests
cryptography==37.0.2
    # via pyjwt
github3-py==3.2.0
    # via -r requirements/prod.in
idna==3.3
    # via requests
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.4.0
    # via github3-py
python-dateutil==2.8.2
    # via github3-py
python-gitlab==3.4.0
    # via -r requirements/prod.in
requests==2.27.1
    # via
    #   github3-py
    #   python-gitlab
    #   requests-toolbelt
requests-toolbelt==0.9.1
    # via python-gitlab
six==1.16.0
    # via python-dateutil
uritemplate==4.1.1
    # via github3-py
urllib3==1.26.9
    # via requests
